Hey Joseph,

On my last import, I noticed a huge performance degradation.  I tracked it down to the salutations extension.  I'll do a "review" on this tomorrow - I'm done with work for the day - but updating custom fields on 100 contacts took this long:
salutations disabled: 11 secs
salutations enabled, 3 salutation types: 47 secs
salutations enabled, 22 salutation types: ~160 secs

I'll go through and explain my changes tomorrow, but I'm getting 27 seconds on 100 contacts (assuming they have 3 salutations each; this version is dependent on the contact's number of salutations, not the total possible # of salutations).

There's two more major optimization I was working on but ran out of steam.  

#### Optimization 1
`hook_civicrm_post` passes through the updated fields as the `$objectRef` argument.  If we can determine that NONE of the tokens on a greeting were changed, we can skip processing the greeting.

#### Optimization 2
An easier win is taking this line in the `Salutation.process` API: 
```php
    $contactDetails = civicrm_api3('Contact', 'getsingle', ['id' => "$contactID",]);
```
and moving into the various functions that call `Salutation.process`.  `Salutation.process` can make the `Contact.get` call only if that param isn't passed.  This will allow us to look up contact details once per contact instead of once per salutation.

I did a synthetic test and found that when we don't have to call `Salutation.process` at all, we get down to 17 seconds, so that's a worthwhile optimization when dealing with imports etc.