@MegaphoneJon 2019-03-28
I just pushed a bug fix to this PR.  The validation rule that prevents saving multiple salutations of the same type was accidentally firing on multi-value custom fields that weren't salutations.  I added a check on the validation to ensure we're actually validating salutations.