@MegaphoneJon 2019-03-29
The latest commit fixes a scenario that gives a false positive for "this salutation already exists" on salutations themselves, as well as simplifies that code a bit.