@MegaphoneJon 2019-09-09
Yes and no?  I've made further changes that I haven't PRed yet, including one this weekend.  I wouldn't trust any but the latest.  I can submit another PR.